// src/types/bodyLanguage.ts
export interface BodyLanguageFeedback {
  postureScore: number;
  postureFeedback: string;
  eyeContactScore: number;
  eyeContactFeedback: string;
  gesturesScore: number;
  gesturesFeedback: string;
  expressionScore: number;
  expressionFeedback: string;
  overallScore: number;
  overallFeedback: string;
  topTips: string[];
}